wclient: use uint8_t to store bufferCount,bufferIndex
Instead of uint32_t as one should never have more
than 255 image buffers allocated at a given time.

+++ include/wclient.h
@@ -232,7 +232,7 @@ kmr_wc_buffer_destroy (struct kmr_wc_buffer *buffer);
  * Given that the arguments of the function are a pointer to a boolean,
  * pointer to an integer, and a pointer to void data type
  */
-typedef void (*kmr_wc_renderer_impl)(volatile bool*, uint32_t*, void*);
+typedef void (*kmr_wc_renderer_impl)(volatile bool*, uint8_t*, void*);
 
 
 /*
@@ -259,7 +259,7 @@ struct kmr_wc_surface {
 	struct xdg_surface          *xdgSurface;
 	struct wl_surface           *wlSurface;
 	struct wl_callback          *wlCallback;
-	uint32_t                    bufferCount;
+	uint8_t                     bufferCount;
 	struct kmr_wc_buffer_handle *wlBufferHandles;
 	void                        *rendererInfo;
 };
@@ -288,12 +288,12 @@ struct kmr_wc_surface {
 struct kmr_wc_surface_create_info {
 	struct kmr_wc_core   *core;
 	struct kmr_wc_buffer *buffer;
-	uint32_t             bufferCount;
+	uint8_t              bufferCount;
 	const char           *appName;
 	bool                 fullscreen;
 	kmr_wc_renderer_impl renderer;
 	void                 *rendererData;
-	uint32_t             *rendererCurrentBuffer;
+	uint8_t              *rendererCurrentBuffer;
 	volatile bool        *rendererRunning;
 };
